backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.33.0
sqlalchemy==1.4.54
pymysql==1.1.1
python-jose[cryptography]==3.4.0
passlib==1.7.4
python-multipart==0.0.20
pydantic==1.10.21
python-dotenv==1.0.1